@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Globalization;
-using System.IO;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using System.Globalization;
 using System.Collections.Concurrent;
 using Eleon.Modding;
 using EmpyrionModWebHost.Extensions;
@@ -12,8 +7,6 @@ using EmpyrionNetAPIAccess;
 using EmpyrionNetAPITools;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using System.Threading;
-using Microsoft.Extensions.Logging;
 using AutoMapper;
 
 namespace EmpyrionModWebHost.Controllers
